pyproject.toml
[tool.poetry]
name = "qa-system"
version = "0.1.0"
description = ""
authors = ["stella"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiohappyeyeballs = "2.4.6"
aiohttp = "3.11.13"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.8.0"
async-timeout = "5.0.1"
attrs = "25.1.0"
bm25s = "0.2.7.post1"
certifi = "2025.1.31"
cfgv = "3.4.0"
charset-normalizer = "3.4.1"
click = "8.1.8"
datasets = "3.3.2"
dill = "0.3.8"
distlib = "0.3.9"
exceptiongroup = "1.2.2"
faiss-cpu = "1.10.0"
fastapi = "0.115.8"
filelock = "3.17.0"
frozenlist = "1.5.0"
fsspec = "2024.12.0"
h11 = "0.14.0"
huggingface-hub = "0.29.1"
identify = "2.6.8"
idna = "3.10"
jinja2 = "3.1.5"
joblib = "1.4.2"
markupsafe = "3.0.2"
mpmath = "1.3.0"
multidict = "6.1.0"
multiprocess = "0.70.16"
networkx = "3.2.1"
nodeenv = "1.9.1"
numpy = "2.0.2"
packaging = "24.2"
pandas = "2.2.3"
pillow = "11.1.0"
platformdirs = "4.3.6"
pre-commit = "4.1.0"
propcache = "0.3.0"
pyarrow = "19.0.1"
pydantic = "2.10.6"
pydantic-core = "2.27.2"
python-dateutil = "2.9.0.post0"
pytz = "2025.1"
pyyaml = "6.0.2"
rank-bm25 = "0.2.2"
regex = "2024.11.6"
requests = "2.32.3"
safetensors = "0.5.3"
scikit-learn = "1.6.1"
scipy = "1.13.1"
sentence-transformers = "3.4.1"
six = "1.17.0"
sniffio = "1.3.1"
starlette = "0.45.3"
sympy = "1.13.1"
threadpoolctl = "3.5.0"
tokenizers = "0.21.0"
torch = "2.6.0"
tqdm = "4.67.1"
transformers = "4.49.0"
typing-extensions = "4.12.2"
tzdata = "2025.1"
urllib3 = "2.3.0"
uvicorn = "0.34.0"
virtualenv = "20.29.2"
wget = "3.2"
xxhash = "3.5.0"
yarl = "1.18.3"
wandb = "^0.19.7"
accelerate = ">=0.26.0"
bitsandbytes = "^0.45.3"
chromadb = "^0.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
